requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
pyyaml>=6.0
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
